split and individualise game page making
